backend/requirements.txt
asgiref==3.4.1
Django==3.2.15
django-filter==22.1
djangorestframework==3.13.1
drf-base64==2.0
fpdf==1.7.2
gunicorn==20.1.0
isort==5.10.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2021.3
reportlab==3.6.11
sqlparse==0.4.2
python-dotenv==0.20.0
djoser==2.1.0
